# Backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: GoProject — Collaborative Note & Team Management API
  description: |
    Full REST API for managing users, teams, folders, notes, and asset sharing.

    ## Authentication
    All protected endpoints require a **Bearer JWT token** in the `Authorization` header.
    Obtain a token via `POST /api/auth/login`. Tokens expire after **24 hours**.

    ## Access Control
    Resources follow a 4-stage check:
    1. **Owner** → full access
    2. **Explicit ACL** (Redis-cached, 30 min TTL)
    3. **Team Manager** of the owner → read-only
    4. **Deny** → HTTP 403

    ## Rate Limits
    | Endpoint | Limit |
    |---|---|
    | `POST /auth/login` | 10,000 req / min / IP |
    | `POST /users/register` | 10,000 req / min / IP |
    | `POST /users/import` | 5 req / min / IP · 3 req / 10 min per user |
  version: 1.0.0
  contact:
    email: duynien34@gmail.com
servers:
  - url: http://localhost:8080/api
    description: Local development
  - url: http://localhost:8080/api
    description: Docker Compose stack
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
...
    CreateTeamRequest:
      type: object
      required: [teamName]
      properties:
        teamName:
          type: string
          example: engineering
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberInput"
    MemberInput:
      type: object
      required: [username, role]
      properties:
        username:
          type: string
          example: bob
        role:
          type: string
          enum: [manager, member]
          example: member
...
    CreateFolderRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: My Project Notes
    UpdateFolderRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Renamed Folder
...
    Team:
      type: object
      properties:
        team_id:
          type: integer
          example: 1
        team_name:
          type: string
          example: engineering
        created_at:
          type: string
          format: date-time
          example: "2026-01-15T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-15T09:00:00Z"
...
    Folder:
      type: object
      properties:
        id:
          type: integer
          example: 7
        owner_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: My Project Notes
        created_at:
          type: string
          format: date-time
          example: "2026-01-15T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-15T09:00:00Z"
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: operation successful
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: invalid request body
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: unauthorized
    Forbidden:
      description: Authenticated but lacks permission on this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: not found
    BadRequest:
      description: Validation error or malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: invalid request body
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: internal server error

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /teams:
    get:
      tags: [Teams]
      summary: List all teams
      description: Returns all teams visible to the authenticated user.
      operationId: listTeams
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Teams retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Teams]
      summary: Create a team
      description: |
        Creates a new team. The authenticated user becomes the `OWNER` automatically.
        Requires `manager` role.
      operationId: createTeam
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
            example:
              teamName: engineering
              members:
                - username: bob
                  role: member
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                type: object
                properties:
                  teamID:
                    type: integer
                    example: 1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /teams/{teamName}:
    delete:
      tags: [Teams]
      summary: Delete a team
      description: |
        Permanently deletes a team and all its memberships.
        Only the team `OWNER` can perform this action.
      operationId: deleteTeam
      security:
        - BearerAuth: []
      parameters:
        - name: teamName
          in: path
          required: true
          schema:
            type: string
          example: engineering
      responses:
        "200":
          description: Team deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Team deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Folders ─────────────────────────────────────────────────────────────────
  /folders:
    post:
      tags: [Folders]
      summary: Create a folder
      description: Creates a new folder owned by the authenticated user.
      operationId: createFolder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [Folders]
      summary: List folders
      description: Returns all folders accessible to the authenticated user (owned + shared).
      operationId: listFolders
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Folders retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  folders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Folder"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /folders/{id}:
    get:
      tags: [Folders]
      summary: Get a folder
      description: |
        Returns a single folder by ID. Access is checked via the 4-stage ACL.
      operationId: getFolder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 7
      responses:
        "200":
          description: Folder found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Folders]
      summary: Update a folder
      description: Renames a folder. Only the owner or a user with `write` permission can update.
      operationId: updateFolder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
      responses:
        "200":
          description: Folder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Folders]
      summary: Delete a folder
      description: |
        Deletes a folder and all its notes. Only the owner can delete.
        Triggers RabbitMQ cache invalidation events for all affected ACL entries.
      operationId: deleteFolder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 7
      responses:
        "200":
          description: Folder deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: folder deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Notes (under folder) ─────────────────────────────────────────────────────
...